Fixed ContactInfoModule (require contactId)
