+++ app/src/main/java/com/zhuravlevmikhail/a65appshomeproject/diContainer/components/AppComponent.kt
@@ -6,7 +6,6 @@ import com.zhuravlevmikhail.a65appshomeproject.diContainer.modules.data.*
 import com.zhuravlevmikhail.a65appshomeproject.diContainer.modules.domain.InteractorsModule
 import com.zhuravlevmikhail.a65appshomeproject.diContainer.modules.domain.ReposModule
 import com.zhuravlevmikhail.a65appshomeproject.diContainer.modules.presentation.ContactInfoModule
-import com.zhuravlevmikhail.a65appshomeproject.fragments.detail.DetailedFragment
 import dagger.Component
 import javax.inject.Singleton
 
@@ -23,6 +22,6 @@ interface AppComponent {
     fun plusContactsComponent() : ContactsComponent
     fun plusDetailedContactComponent() : DetailedContactComponent
     fun plusMainComponent() : ActivityHostComponent
-    fun plusMapComponent(): MapComponent
+    fun plusMapComponent(contactInfoModule: ContactInfoModule): ContactInfoComponent
     fun inject(app : App)
 }
