+++ app/src/main/java/com/zhuravlevmikhail/a65appshomeproject/diContainer/modules/presentation/PresentersModule.kt
@@ -1,16 +1,11 @@
 package com.zhuravlevmikhail.a65appshomeproject.diContainer.modules.presentation
 
-import android.content.Context
-import com.zhuravlevmikhail.a65appshomeproject.R
 import com.zhuravlevmikhail.a65appshomeproject.domain.contacts.ContactsInteractor
 import com.zhuravlevmikhail.a65appshomeproject.diContainer.scopes.FragmentScope
-import com.zhuravlevmikhail.a65appshomeproject.domain.map.MapInteractor
 import com.zhuravlevmikhail.a65appshomeproject.fragments.contacts.ContactsPresenter
 import com.zhuravlevmikhail.a65appshomeproject.fragments.detail.DetailedPresenter
-import com.zhuravlevmikhail.a65appshomeproject.fragments.detail.innerFragments.MapPresenter
 import dagger.Module
 import dagger.Provides
-import javax.inject.Named
 
 @Module
 class PresentersModule {
@@ -25,16 +20,5 @@ class PresentersModule {
     @FragmentScope
     fun provideDetailedPresenter(contactsInteractor: ContactsInteractor) : DetailedPresenter {
         return DetailedPresenter(contactsInteractor)
-    }
-
-    @Provides
-    @FragmentScope
-    fun provideMapPresenter(mapInteractor: MapInteractor, context: Context, @Named("contactId") contactId : Long): MapPresenter {
-        val apiKey = context.resources.getString(R.string.yandex_api_key)
-        return MapPresenter(
-            mapInteractor,
-            apiKey,
-            contactId
-        )
     }
 }
